refactor(server): isolate platform driver config
